Move kr8s_cluster fixture to root level
